fix bug where sending create/delete request twice
this was because the code that let the rest of the changes know that we
were back online was writing the same change that we received over the
wire that we just sent. it created 2 processes that were simultaneously
executing the requests.

the logic was updated so that the start of each loop will process each
requests one at a time. looking at it now, i think there's still a race
here and we should be invalidating the entire array and setting it back
at the end

update - fixed the todo from above in an ammend

fixes #5